Move Search component
